# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/health": {
      "get": {
        "operationId": "HealthController_health",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Check API and database health",
        "tags": [
          "health"
        ]
      }
    },
    "/api/v1/analysis/website": {
      "post": {
        "operationId": "AnalysisController_analyzeWebsite",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Analyze a public company website with SSRF protection",
        "tags": [
          "analysis"
        ]
      }
    },
    "/api/v1/dashboard": {
      "get": {
        "operationId": "DashboardController_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get the current workspace pipeline summary",
        "tags": [
          "dashboard"
        ]
      }
    },
    "/api/v1/companies": {
      "get": {
        "operationId": "CompaniesController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List workspace companies",
        "tags": [
          "companies"
        ]
      },
      "post": {
        "operationId": "CompaniesController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a company",
        "tags": [
          "companies"
        ]
      }
    },
    "/api/v1/companies/{id}": {
      "get": {
        "operationId": "CompaniesController_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a company",
        "tags": [
          "companies"
        ]
      },
      "patch": {
        "operationId": "CompaniesController_update",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a company",
        "tags": [
          "companies"
        ]
      },
      "delete": {
        "operationId": "CompaniesController_delete",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a company",
        "tags": [
          "companies"
        ]
      }
    },
    "/api/v1/companies/{id}/analyze": {
      "post": {
        "operationId": "CompaniesController_analyze",
        "parameters": [],
        "responses": {
          "202": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Queue analysis for a company website and public profile",
        "tags": [
          "companies"
        ]
      }
    },
    "/api/v1/contacts": {
      "get": {
        "operationId": "ContactsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List workspace contacts",
        "tags": [
          "contacts"
        ]
      },
      "post": {
        "operationId": "ContactsController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a contact",
        "tags": [
          "contacts"
        ]
      }
    },
    "/api/v1/contacts/{id}": {
      "get": {
        "operationId": "ContactsController_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a contact",
        "tags": [
          "contacts"
        ]
      },
      "patch": {
        "operationId": "ContactsController_update",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a contact",
        "tags": [
          "contacts"
        ]
      },
      "delete": {
        "operationId": "ContactsController_delete",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a contact",
        "tags": [
          "contacts"
        ]
      }
    },
    "/api/v1/leads": {
      "get": {
        "operationId": "LeadsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List and filter workspace leads",
        "tags": [
          "leads"
        ]
      },
      "post": {
        "operationId": "LeadsController_create",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a manual or pre-qualified lead",
        "tags": [
          "leads"
        ]
      }
    },
    "/api/v1/leads/{id}": {
      "get": {
        "operationId": "LeadsController_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a lead with company, contact, signals, score, and activity",
        "tags": [
          "leads"
        ]
      },
      "patch": {
        "operationId": "LeadsController_update",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a lead",
        "tags": [
          "leads"
        ]
      },
      "delete": {
        "operationId": "LeadsController_delete",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a lead",
        "tags": [
          "leads"
        ]
      }
    },
    "/api/v1/leads/{id}/scores": {
      "get": {
        "operationId": "LeadsController_scores",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get score history for a lead",
        "tags": [
          "leads"
        ]
      }
    },
    "/api/v1/leads/{id}/analyze": {
      "post": {
        "operationId": "LeadsController_analyze",
        "parameters": [],
        "responses": {
          "202": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Queue AI/manual lead analysis",
        "tags": [
          "leads"
        ]
      }
    },
    "/api/v1/leads/{id}/rescore": {
      "post": {
        "operationId": "LeadsController_rescore",
        "parameters": [],
        "responses": {
          "202": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Queue a lead rescore using current settings",
        "tags": [
          "leads"
        ]
      }
    },
    "/api/v1/leads/{id}/score-override": {
      "post": {
        "operationId": "LeadsController_overrideScore",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Manually override the latest visible lead score without deleting history",
        "tags": [
          "leads"
        ]
      }
    },
    "/api/v1/jobs": {
      "get": {
        "operationId": "JobsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List workspace system jobs",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/jobs/{id}": {
      "get": {
        "operationId": "JobsController_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a workspace system job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/v1/settings": {
      "get": {
        "operationId": "SettingsController_get",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get editable workspace settings",
        "tags": [
          "settings"
        ]
      },
      "patch": {
        "operationId": "SettingsController_update",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update workspace settings",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/v1/settings/scoring": {
      "get": {
        "operationId": "SettingsController_getScoring",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get deterministic lead scoring weights",
        "tags": [
          "settings"
        ]
      },
      "patch": {
        "operationId": "SettingsController_updateScoring",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update deterministic lead scoring weights",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/v1/settings/ai": {
      "get": {
        "operationId": "SettingsController_getAiStatus",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get OpenAI configuration status without exposing secrets",
        "tags": [
          "settings"
        ]
      }
    }
  },
  "info": {
    "title": "Scrapo Lead Hunter API",
    "description": "Independent, workspace-scoped REST API for the Scrapo Lead Hunter clients.",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "workspace": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Workspace-Id"
      }
    },
    "schemas": {}
  }
}